moving namespaces, adding constants to avoid magic strings, ResultFile data helpers
