+++ lib/Hutch.Rackit/TaskApi/Models/AvailabilityJob.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Hutch.Rackit.Models.TaskApi;
+namespace Hutch.Rackit.TaskApi.Models;
 
 public class AvailabilityJob : TaskApiBaseResponse
 {
